fix: Manually count indexeddb size rather than using FetchUpdates()
FetchUpdates() emits the 'synced' event after loading content into the
ydoc, but before counting the number of row. This means that
persistence._dbsize is still 0 for event listeners to once("synced",
...). We used to call FetchUpdates again to ensure that the count
happened, but I noticed that this would cause all updates to be loaded
twice, and could trigger a lot of on("update", ...) events.

We also now delay startup until the document level whenSynced() event is
fired, and we move the manual count() into this gate. It would probably
be cleaner to vendor and modify y-indexeddb.

OTOH newer version of the y-sweet provider actually have encrypted
persistence built in, which is pretty compelling. We've also considered
writing our own persistence to store files on disk.